indoxRouter_dashboard/requirements.txt
streamlit>=1.31.0
psycopg2-binary>=2.9.9
pymongo[srv]>=4.6.1
dnspython>=2.4.0
httpx>=0.24.0
python-dotenv>=1.0.0
bcrypt>=4.0.1
python-jose>=3.3.0
pydantic>=2.0.0
faker>=22.0.0
uuid>=1.30
plotly>=5.18.0
pandas>=2.0.0 